+++ genienlp/arguments.py
@@ -308,7 +308,11 @@ def parse_argv(parser):
         help='alternative momentum for Adam (only when not using transformer scheduler), and RAdam',
     )
     parser.add_argument(
-        '--optimizer', default='adam', choices=['adam', 'adamw', 'sgd', 'radam'], type=str, help='optimizer to use'
+        '--optimizer',
+        default='adam',
+        choices=['adam', 'adamw', 'adafactor', 'radam', 'sgd'],
+        type=str,
+        help='optimizer to use',
     )
     parser.add_argument(
         '--lr_schedule',
